# api/openapi.yaml
openapi: 3.0.3
info:
  title: Mayastor RESTful API
  version: v0
servers:
- url: /v0
paths:
...
  /nodes/{node_id}/volumes:
    get:
      operationId: get_node_volumes
      parameters:
      - explode: false
        in: path
        name: node_id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Volume'
                type: array
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Request Timeout
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Not Found
        "408":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Bad Request
        "412":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Precondition Failed
        "416":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Range Not satisfiable
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Unprocessable entity
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Not Implemented
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Service Unavailable
        "504":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Gateway Timeout
        "507":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Insufficient Storage
      security:
      - JWT: []
      tags:
      - Volumes
  /nodes/{node_id}/volumes/{volume_id}:
    get:
      operationId: get_node_volume
      parameters:
      - explode: false
        in: path
        name: node_id
        required: true
        schema:
          type: string
        style: simple
      - explode: false
        in: path
        name: volume_id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Volume'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Request Timeout
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Not Found
        "408":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Bad Request
        "412":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Precondition Failed
        "416":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Range Not satisfiable
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Unprocessable entity
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Not Implemented
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Service Unavailable
        "504":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Gateway Timeout
        "507":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Insufficient Storage
      security:
      - JWT: []
      tags:
      - Volumes
...
  /volumes:
    get:
      operationId: get_volumes
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Volume'
                type: array
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Request Timeout
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Unauthorized
        "408":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Bad Request
        "412":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Precondition Failed
        "416":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Range Not satisfiable
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Unprocessable entity
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Not Implemented
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Service Unavailable
        "504":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Gateway Timeout
        "507":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Insufficient Storage
      security:
      - JWT: []
      tags:
      - Volumes
  /volumes/{volume_id}:
    delete:
      operationId: del_volume
      parameters:
      - explode: false
        in: path
        name: volume_id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "204":
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Request Timeout
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Not Found
        "408":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Bad Request
        "412":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Precondition Failed
        "416":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Range Not satisfiable
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Unprocessable entity
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Not Implemented
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Service Unavailable
        "504":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Gateway Timeout
        "507":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Insufficient Storage
      security:
      - JWT: []
      tags:
      - Volumes
    get:
      operationId: get_volume
      parameters:
      - explode: false
        in: path
        name: volume_id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Volume'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Request Timeout
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Not Found
        "408":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Bad Request
        "412":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Precondition Failed
        "416":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Range Not satisfiable
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Unprocessable entity
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Not Implemented
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Service Unavailable
        "504":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Gateway Timeout
        "507":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Insufficient Storage
      security:
      - JWT: []
      tags:
      - Volumes
    put:
      operationId: put_volume
      parameters:
      - explode: false
        in: path
        name: volume_id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateVolumeBody'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Volume'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Request Timeout
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Not Found
        "408":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Bad Request
        "412":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Precondition Failed
        "416":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Range Not satisfiable
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Unprocessable entity
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Not Implemented
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Service Unavailable
        "504":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Gateway Timeout
        "507":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Insufficient Storage
      security:
      - JWT: []
      tags:
      - Volumes
  /volumes/{volume_id}/share/{protocol}:
    put:
      operationId: put_volume_share
      parameters:
      - explode: false
        in: path
        name: volume_id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      - explode: false
        in: path
        name: protocol
        required: true
        schema:
          enum:
          - nvmf
          - iscsi
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                type: string
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Request Timeout
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Not Found
        "408":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Bad Request
        "412":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Precondition Failed
        "416":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Range Not satisfiable
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Unprocessable entity
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Not Implemented
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Service Unavailable
        "504":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Gateway Timeout
        "507":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Insufficient Storage
      security:
      - JWT: []
      tags:
      - Volumes
  /volumes{volume_id}/share:
    delete:
      operationId: del_shares
      parameters:
      - explode: false
        in: path
        name: volume_id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "204":
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Request Timeout
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Not Found
        "408":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Bad Request
        "412":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Precondition Failed
        "416":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Range Not satisfiable
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Unprocessable entity
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Not Implemented
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Service Unavailable
        "504":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Gateway Timeout
        "507":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Insufficient Storage
      security:
      - JWT: []
      tags:
      - Volumes
...
components:
  schemas:
...
    CreateVolumeBody:
      description: Create Volume Body JSON
      example:
        policy:
          self_heal: false
          topology: null
        replicas: 0
        size: 0
        topology:
          explicit: null
          labelled: null
      properties:
        policy:
          $ref: '#/components/schemas/CreateVolumeBody_policy'
        replicas:
          description: number of storage replicas
          format: int64
          type: integer
        size:
          description: size of the volume in bytes
          format: int64
          type: integer
        topology:
          $ref: '#/components/schemas/CreateVolumeBody_topology'
      required:
      - policy
      - replicas
      - size
      - topology
      type: object
...
    RestJsonError:
      description: Rest Json Error format
      example:
        details: ""
        error: NotFound
      properties:
        details:
          description: detailed error information
          type: string
        error:
          description: error kind
          enum:
          - Timeout
          - Deserialize
          - Internal
          - InvalidArgument
          - DeadlineExceeded
          - NotFound
          - AlreadyExists
          - PermissionDenied
          - ResourceExhausted
          - FailedPrecondition
          - NotShared
          - NotPublished
          - AlreadyPublished
          - AlreadyShared
          - Aborted
          - OutOfRange
          - Unimplemented
          - Unavailable
          - Unauthenticated
          - Unauthorized
          - Conflict
          - FailedPersist
          - Deleting
          type: string
      required:
      - details
      - error
      type: object
...
    Volume:
      description: |-
        Volumes

         Volume information
      example:
        children:
        - children:
          - rebuildProgress: null
            state: Unknown
            uri: ""
          deviceUri: ""
          node: ""
          rebuilds: 0
          share: "off"
          size: 0
          state: Unknown
          uuid: 514ed1c8-7174-49ac-b9cd-ad44ef670a67
        protocol: "off"
        size: 0
        state: Unknown
        uuid: 514ed1c8-7174-49ac-b9cd-ad44ef670a67
      properties:
        children:
          description: array of children nexuses
          items:
            $ref: '#/components/schemas/Volume_children'
          type: array
        protocol:
          description: current share protocol
          enum:
          - "false"
          - nvmf
          - iscsi
          - nbd
          type: string
        size:
          description: size of the volume in bytes
          format: int64
          type: integer
        state:
          description: current state of the volume
          enum:
          - Unknown
          - Online
          - Degraded
          - Faulted
          type: string
        uuid:
          description: name of the volume
          format: uuid
          type: string
      required:
      - children
      - protocol
      - size
      - state
      - uuid
      type: object
...
    CreateVolumeBody_policy_topology_explicit:
      description: volume topology, explicitly selected
      example:
        allowed_nodes:
        - ""
        preferred_nodes:
        - ""
      properties:
        allowed_nodes:
          description: replicas can only be placed on these nodes
          items:
            type: string
          type: array
        preferred_nodes:
          description: preferred nodes to place the replicas
          items:
            type: string
          type: array
      required:
      - allowed_nodes
      - preferred_nodes
      type: object
...
    CreateVolumeBody_policy_topology_labelled_pool_topology:
      description: pool topology
      example:
        inclusion:
        - ""
      properties:
        inclusion:
          description: inclusive labels
          items:
            description: |-
              Includes resources with the same $label or $label:$value eg:
               if label is "Zone: A":
               A resource with "Zone: A" would be paired up with a resource with "Zone: A",
               but not with a resource with "Zone: B"
               if label is "Zone":
               A resource with "Zone: A" would be paired up with a resource with "Zone: B",
               but not with a resource with "OtherLabel: B"
               inclusive label key value in the form "NAME: VALUE"
            example: ""
            type: string
          type: array
      required:
      - inclusion
      type: object
    CreateVolumeBody_policy_topology_labelled:
      description: volume topology using labels
      example:
        node_topology:
          exclusion:
          - ""
          inclusion:
          - ""
        pool_topology:
          inclusion:
          - ""
      properties:
        node_topology:
          $ref: '#/components/schemas/CreateVolumeBody_policy_topology_labelled_node_topology'
        pool_topology:
          $ref: '#/components/schemas/CreateVolumeBody_policy_topology_labelled_pool_topology'
      required:
      - node_topology
      - pool_topology
      type: object
    CreateVolumeBody_policy_topology:
      description: |-
        topology to choose a replacement replica for self healing
         (overrides the initial creation topology)
      example:
        explicit: null
        labelled: null
      properties:
        explicit:
          $ref: '#/components/schemas/CreateVolumeBody_policy_topology_explicit'
        labelled:
          $ref: '#/components/schemas/CreateVolumeBody_policy_topology_labelled'
      type: object
    CreateVolumeBody_policy:
      description: Volume Healing policy used to determine if and how to replace a
        replica
      example:
        self_heal: false
        topology: null
      properties:
        self_heal:
          description: |-
            the server will attempt to heal the volume by itself
             the client should not attempt to do the same if this is enabled
          type: boolean
        topology:
          $ref: '#/components/schemas/CreateVolumeBody_policy_topology'
      required:
      - self_heal
      type: object
    CreateVolumeBody_topology:
      description: |-
        Volume topology used to determine how to place/distribute the data
         Should either be labelled or explicit, not both.
         If neither is used then the control plane will select from all available resources.
      example:
        explicit: null
        labelled: null
      properties:
        explicit:
          $ref: '#/components/schemas/CreateVolumeBody_policy_topology_explicit'
        labelled:
          $ref: '#/components/schemas/CreateVolumeBody_policy_topology_labelled'
      type: object
    Nexus_children:
      description: Child information
      example:
        rebuildProgress: null
        state: Unknown
        uri: ""
      properties:
        rebuildProgress:
          description: current rebuild progress (%)
          format: int32
          type: integer
        state:
          description: state of the child
          enum:
          - Unknown
          - Online
          - Degraded
          - Faulted
          type: string
        uri:
          description: uri of the child device
          type: string
      required:
      - state
      - uri
      type: object
...
    Volume_children:
      description: Nexus information
      example:
        children:
        - rebuildProgress: null
          state: Unknown
          uri: ""
        deviceUri: ""
        node: ""
        rebuilds: 0
        share: "off"
        size: 0
        state: Unknown
        uuid: 514ed1c8-7174-49ac-b9cd-ad44ef670a67
      properties:
        children:
          description: array of children
          items:
            $ref: '#/components/schemas/Nexus_children'
          type: array
        deviceUri:
          description: |-
            URI of the device for the volume (missing if not published).
             Missing property and empty string are treated the same.
          type: string
        node:
          description: id of the mayastor instance
          type: string
        rebuilds:
          description: total number of rebuild tasks
          format: int32
          type: integer
        share:
          description: protocol used for exposing the nexus
          enum:
          - "false"
          - nvmf
          - iscsi
          - nbd
          type: string
        size:
          description: size of the volume in bytes
          format: int64
          type: integer
        state:
          description: current state of the nexus
          enum:
          - Unknown
          - Online
          - Degraded
          - Faulted
          type: string
        uuid:
          description: uuid of the nexus
          format: uuid
          type: string
      required:
      - children
      - deviceUri
      - node
      - rebuilds
      - share
      - size
      - state
      - uuid
      type: object
  securitySchemes:
    JWT:
      bearerFormat: JWT
      scheme: bearer
      type: http